Fix links in README file
Add a closing bracket (']') at the end of the Coding Challenge #32.1
video's heading ID, right before the full URL of the video,
encapsulated within parentheses. Without said symbol, the markdown
syntax was displayed.

Agar.io's web version link pointed to the game's code. Replace the
destination URL with a link to the sketch running in a fullscreen
browser window.